+++ source/Tubeshade.Server/HumanReadablePeriodPattern.cs
@@ -1,20 +1,61 @@
 ﻿using System;
 using System.Globalization;
 using System.Text;
+using System.Text.RegularExpressions;
 using NodaTime;
 using NodaTime.Text;
 
 namespace Tubeshade.Server;
 
-public sealed class HumanReadablePeriodPattern : IPattern<Period>
+public sealed partial class HumanReadablePeriodPattern : IPattern<Period>
 {
     public static readonly HumanReadablePeriodPattern Instance = new();
 
     /// <inheritdoc />
     public ParseResult<Period> Parse(string text)
     {
-        throw new NotSupportedException("Cannot parse human readable periods");
+        var match = ParseRegex().Match(text);
+        if (!match.Success ||
+            match.Groups is [_, { Length: <= 1 }, { Length: <= 1 }, { Length: <= 1 }, { Length: <= 1 }])
+        {
+            return ParseResult<Period>.ForException(() => new FormatException("Could not parse period"));
+        }
+
+        var builder = new PeriodBuilder();
+
+        try
+        {
+            if (match.Groups[1].ValueSpan is { Length: > 1 } days)
+            {
+                builder.Days = int.Parse(days[..^1], CultureInfo.InvariantCulture);
+            }
+
+            if (match.Groups[2].ValueSpan is { Length: > 1 } hours)
+            {
+                builder.Hours = int.Parse(hours[..^1], CultureInfo.InvariantCulture);
+            }
+
+            if (match.Groups[3].ValueSpan is { Length: > 1 } minutes)
+            {
+                builder.Minutes = int.Parse(minutes[..^1], CultureInfo.InvariantCulture);
+            }
+
+            if (match.Groups[4].ValueSpan is { Length: > 1 } seconds)
+            {
+                builder.Seconds = int.Parse(seconds[..^1], CultureInfo.InvariantCulture);
+            }
+        }
+        catch (Exception exception)
+        {
+            return ParseResult<Period>.ForException(() => exception);
+        }
+
+
+        return ParseResult<Period>.ForValue(builder.Build());
     }
+
+    [GeneratedRegex(@"^\s*(\d{1,}d)?\s*(\d{1,}h)?\s*(\d{1,}m)?\s*(\d{1,}s)?\s*$")]
+    private static partial Regex ParseRegex();
 
     /// <inheritdoc />
     public string Format(Period value) => AppendFormat(value, new StringBuilder()).ToString();